Implemented I2C bus file descriptor reuse.
The first time an I2C bus object for a particular bus is created or
initialized, the file descriptor will be saved and a reference counter
for it initialized to 1.

Each time another I2C bus object for the same bus is created or initialized,
the previously saved file descriptor will be reused and the reference
counter incremented.

Each time an I2C bus object for the same bus is destroyed, the reference
counter for that file descriptor will be decremented.  When the reference
counter reaches zero, the file descriptor will be closed.

The file descriptors and reference counters are embedded within a protected
object, so everything works even if different threads attempt to create or
destroy I2C bus objects at the same time.


git-svn-id: svn+ssh://subversion.munts.net/src/trunk/libsimpleio@20926 06ee2f96-5407-11de-8bab-f3d7589a4122